% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmice_functions.R
\name{call_mice}
\alias{call_mice}
\title{Calls mice's imputation function with parameters provided in a list 'params'}
\usage{
call_mice(params, data_uploads)
}
\arguments{
\item{params}{A list with elements: `data`: A string with the name of the
dataset, a hash from an uploaded file, or a csv file; `maxit`:
number of iterations; `m`: number of imputations, `seed`: seed.}

\item{data_uploads}{path to uploaded data on server}
}
\description{
Calls mice's imputation function with parameters provided in a list 'params'
}
\seealso{
\link[mice]{mice}
}
