% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmice_functions.R
\name{example_data_to_json}
\alias{example_data_to_json}
\title{Fetches example data from mice, returns data as json}
\usage{
example_data_to_json(name)
}
\arguments{
\item{name}{An object name (given as a character string or a symbol).}
}
\value{
Data as json
}
\description{
Fetches example data from mice, returns data as json
}
\seealso{
\link[base]{get}
}
