% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmice_functions.R
\name{imp_result_long_fmt}
\alias{imp_result_long_fmt}
\title{Takes the result of the imputation and returns the long format of the data}
\usage{
imp_result_long_fmt(imp)
}
\arguments{
\item{imp}{Multiply imputed data set, an object of class \link[mice]{mids}}
}
\description{
Takes the result of the imputation and returns the long format of the data
}
\seealso{
\link[mice]{complete}
}
