% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmice_functions.R
\name{impute}
\alias{impute}
\title{Mice functions}
\usage{
impute(data, maxit, m, seed)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.  Missing
values are coded as \code{NA}.}

\item{maxit}{A scalar giving the number of iterations. The default is 5.}

\item{m}{Number of multiple imputations. The default is \code{m=5}.}

\item{seed}{An integer that is used as argument by the \code{set.seed()} for
offsetting the random number generator. Default is to leave the random number
generator alone.}
}
\description{
Mice functions
}
