% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmice_functions.R
\name{json_to_parameters}
\alias{json_to_parameters}
\title{Takes a json string and returns it as R list}
\usage{
json_to_parameters(json_payload)
}
\arguments{
\item{json_payload}{The json payload}
}
\description{
Takes a json string and returns it as R list
}
\examples{
input <- jsonlite::toJSON(list(data="nhanes", maxit=2, m=2, seed=1))
json_to_parameters(input)
}
