% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmice_handlers.R
\name{pool_handler}
\alias{pool_handler}
\alias{fit_handler}
\alias{impute_longfmt_handler}
\alias{example_data_handler}
\alias{mice_version_handler}
\title{End point handlers}
\usage{
pool_handler(.req, .res)

fit_handler(.req, .res)

impute_longfmt_handler(.req, .res, data_uploads)

example_data_handler(.req, .res)

mice_version_handler(.req, .res)
}
\arguments{
\item{.req}{Required arguments}

\item{.res}{Result}

\item{data_uploads}{Server directory with uploaded data}
}
\description{
End point handlers
}
